# docs/swagger.yaml
definitions:
  api.AuthRes:
    properties:
      exchanger_balance:
        type: string
      exchanger_flag:
        type: boolean
      status:
        description: 2 The payment status of the exchange is normal, other numbers
          are arrears or no payment
        type: integer
    type: object
...
  api.Subscribe:
    properties:
      email:
        type: string
    type: object
...
  model.Exchanger:
    properties:
      address:
        description: Exchange address
        type: string
      balance_count:
        description: Total transaction amount, unit wei
        type: string
      block_number:
        description: The block number when created
        type: integer
      creator:
        description: Creator address
        type: string
      fee_ratio:
        description: fee rate, unit 1/10,000
        type: integer
      name:
        description: Exchange name
        type: string
      nft_count:
        description: Total NFT count
        type: integer
      timestamp:
        description: Open time
        type: integer
      tx_hash:
        description: The transaction created
        type: string
      url:
        description: Exchange URL
        type: string
    type: object
...
  model.Subscription:
    properties:
      email:
        description: Email
        type: string
    type: object
...
  service.ErrRes:
    properties:
      err_str:
        description: Error message
        type: string
    type: object
  service.ExchangerRes:
    properties:
      address:
        description: Exchange address
        type: string
      amount:
        description: pledge amount
        type: string
      balance_count:
        description: Total transaction amount, unit wei
        type: string
      block_number:
        description: The block number when created
        type: integer
      collectionCount:
        description: collection count
        type: integer
      creator:
        description: Creator address
        type: string
      fee_ratio:
        description: fee rate, unit 1/10,000
        type: integer
      name:
        description: Exchange name
        type: string
      nft_count:
        description: Total NFT count
        type: integer
      timestamp:
        description: Open time
        type: integer
      tx_hash:
        description: The transaction created
        type: string
      txCount:
        description: transaction count
        type: integer
      url:
        description: Exchange URL
        type: string
    type: object
  service.ExchangerTxCountRes:
    properties:
      day_1:
        type: integer
      day_7:
        type: integer
      day30:
        type: integer
      total:
        type: integer
    type: object
  service.ExchangersRes:
    properties:
      exchangers:
        description: List of exchanges
        items:
          $ref: '#/definitions/model.Exchanger'
        type: array
      last_0_total:
        description: The number of newly opened exchanges in the latest 0 days (today),
          calculated in real time
        type: integer
      last_1_total:
        description: Number of newly opened exchanges in the latest 1 day (yesterday),
          cached
        type: integer
      last_7_total:
        description: Number of newly opened exchanges in the last 7 days, cached
        type: integer
      total:
        description: Total number of exchanges
        type: integer
    type: object
...
info:
  contact: {}
  description: Block browser back-end interface, parses data from the blockchain,
    provides information retrieval services for blocks, transactions, NFT, SNFT, NFT
    collections, and exchanges
  title: block explorer API
  version: "1.0"
paths:
...
  /erb_faucet:
    get:
      consumes:
      - application/json
      description: request ERB test coins
      parameters:
      - description: address
        in: query
        name: addr
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: request ERB test coins
      tags:
      - other interfaces
...
  /exchanger/{addr}:
    get:
      consumes:
      - application/json
      description: Query exchanges by address
      parameters:
      - description: Exchange address
        in: path
        name: addr
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ExchangerRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query exchange
      tags:
      - Exchange
...
  /exchanger/page:
    get:
      consumes:
      - application/json
      description: Query the list of exchanges in reverse order of creation time
      parameters:
      - description: Exchange name, if empty, query all exchanges
        in: query
        name: name
        type: string
      - description: Page, default 1
        in: query
        name: page
        type: string
      - description: Page size, default 10
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ExchangersRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: Query the list of exchanges
      tags:
      - Exchange
  /exchanger/tx_count/{addr}:
    get:
      consumes:
      - application/json
      description: Query the exchanges tx count chart
      parameters:
      - description: Exchanger address
        in: path
        name: addr
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.ExchangerTxCountRes'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: Query the exchanges tx count
      tags:
      - Exchange
  /exchanger_auth:
    get:
      consumes:
      - application/json
      description: query exchange status
      parameters:
      - description: address
        in: query
        name: addr
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AuthRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query exchange status
      tags:
      - other interfaces
  /exchangers:
    get:
      consumes:
      - application/json
      description: Query the list of exchanges in reverse order of creation time
      parameters:
      - description: Page, default 1
        in: query
        name: page
        type: string
      - description: Page size, default 10
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.ExchangerRes'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: Query the list of exchanges
      tags:
      - Exchange
...
  /subscription:
    get:
      consumes:
      - application/json
      description: Query the list of subscription mailboxes
      parameters:
      - description: Page, default 1
        in: query
        name: page
        type: string
      - description: Page size, default 10
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: Query the list of subscription mailboxes
      tags:
      - other interfaces
    post:
      consumes:
      - application/json
      description: Enter the email address to receive the latest event notifications
      parameters:
      - description: Mailbox
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/api.Subscribe'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: subscribe email
      tags:
      - other interfaces
...
swagger: "2.0"